+++ galacteek/ld/__init__.py
@@ -1,0 +1,2 @@
+def ipsContextUri(contextName):
+    return f'ips://galacteek.ld.contexts/{contextName}'